Checksumming of IPv4 header added as an option
The funtion was written and tested for IPv4, but it is commented
out, as checksumming problems seems to be an issue of TCP and UDP
not IPv4.

This issue appears in TAP devices where kernel keeps track of the
checksum status, but this program does not know that checksums were
not computed.

ethtool -K dev tx off

does the trick and it forces kernel to compute outgoing checksums.
tap offloading = do nothing, unlike hardware devices.