Fixing and debugging MSG_DONTWAIT
As it turns out, MSG_DONTWAIT is not asynchronous write. It returns
either -1 or number of bytes written which might be less than scheduled.
One has to keep track of the number of bytes written and the number of
bytes pending and if leftovers are to be sent, call send() again.