A crash was observed, adding a few extra prints
The crash was reported with "Dupa 33" message and an info,
that entered reading while not all writes were completed.

It appears a blocking send() might return nonzero status and
return a positive value of bytes written, which is less than
bytes pending. Added logic to take this possibility into account.

This probably happens when a netdev goes down.